Add new repo *pia* (#49)
* Add new repo *pia*

*pia*, short for Project Identity Authority, will provide an
oidc-based authentication broker for Eclipse Foundation projects.

The motivating use case is to authenticate SBOM uploads to
https://sbom.eclipse.org from Jenkins CI and GitHub Actions.
See https://docs.google.com/document/d/1DuvC7vRsbJQpY9q4selLuAuAO4AUXiyEmHrK-xQ6GW0
for a problem statement and basic design.

An end-to-end POC can be found in
https://gitlab.eclipse.org/lukpueh/oidc-upload-demo. The "broker"
component in the demo is the prototype for PIA.

Alternative naming suggestions:
* brokerDog (wordplay on eclipse-csi sister project otterDog)
* handover
* sink
* up
* anvil (metaphor for forging identity)
* `<your naming suggestion here>`

Signed-off-by: Lukas Puehringer <lukas.puehringer@eclipse-foundation.org>

* Fix jsonnet syntax for new repo *pia*

Signed-off-by: Lukas Puehringer <lukas.puehringer@eclipse-foundation.org>

---------

Signed-off-by: Lukas Puehringer <lukas.puehringer@eclipse-foundation.org>